Cleaned up repo
Instead of storing an entire Visual Studio project, the repository only
keeps the necessary header files, source files, and libraries for
building a project from scratch with any code development suite (e.g.,
could use Xcode on Mac). In addition, the repo holds a “Notes” folder,
which contains any unintuitive information regarding the code. In the
future, it should have an updated how-to guide for creating a fresh
Visual Studio project (an old version of this is on the desktop of the
BrainGate computer).